hiring-de-geolocation-data-main/requirements.txt
pyspark
geopy
apache-airflow
pytest